finish part 4. Except for exercises 4.6, 4.7. Lecture tests not fixed.

+++ part4/blog/controllers/blogs.js
@@ -1,25 +1,55 @@
+const jwt = require('jsonwebtoken');
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
+const User = require('../models/user');
 
 blogsRouter.get('/', async (req, res) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 });
   res.json(blogs);
 });
 
 blogsRouter.post('/', async (req, res) => {
-  if (!('likes' in req.body)) {
+  const body = req.body;
+
+  if (!('likes' in body)) {
     req.body['likes'] = 0;
   }
 
-  const blog = new Blog(req.body);
+  const token = req.token;
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!token || !decodedToken.id) {
+    return res.status(401).json({ error: 'token missing or invalid' });
+  }
+
+  const user = await User.findById(decodedToken.id);
+
+  const blog = new Blog({ ...body, user: user._id });
 
   const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
   res.status(201).json(savedBlog);
 });
 
 blogsRouter.delete('/:id', async (req, res) => {
-  await Blog.findByIdAndRemove(req.params.id);
-  res.status(204).end();
+  const id = req.params.id;
+
+  const token = req.token;
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!token || !decodedToken.id) {
+    return res.status(401).json({ error: 'token missing or invalid' });
+  }
+
+  const user = await User.findById(decodedToken.id);
+  const blog = await Blog.findById(id);
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(req.params.id);
+    res.status(204).end();
+  } else {
+    res.status(401).json({ error: 'cannot delete blog of another user' });
+  }
 });
 
 blogsRouter.put('/:id', async (req, res) => {
